Factory Issue is Handled
We could copy the improvement script and keep lots of things in that script, so we did it.
The most important point of that commit is we have to check the amount of raw materials whether it is enough to process raw materials to make processed materials which is the main idea of those factories. To sum up, that's what we did.